add cast to fix discount price

+++ bookeshop-data-model/src/main/java/org/benassi/bookeshop/data/model/Book.java
@@ -137,7 +137,7 @@ public class Book {
     private String stockStatus;
 
     public float getDiscountPrice(){
-        return price * ( 1 - offer / RATIO );
+        return price * ( 1 - (float) offer / RATIO );
     }
 
     public String getFormattedPublishDate() {
